backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgi-lifespan==2.1.0
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
dnspython==2.6.1
docker==7.0.0
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
